--- scripts/mbedtls_dev/ecp_common.py
@@ -1,22 +1,0 @@
-"""Common features for ecp in test generation framework."""
-# Copyright The Mbed TLS Contributors
-# SPDX-License-Identifier: Apache-2.0
-#
-# Licensed under the Apache License, Version 2.0 (the "License"); you may
-# not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
-# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from . import bignum_common
-
-class EcpOperationCommon(bignum_common.ModOperationCommon):
-    #pylint: disable=abstract-method
-    """Target for ecp test case generation."""
-    pass
